requirements.txt
amqp==5.0.6
celery==5.1.0
channels==3.0.3
channels-redis==3.2.0
Django==3.1.4
djongo==1.3.6
django-annoying==0.10.6
django-minio-storage==0.3.10
django-cors-headers==3.7.0
django-notifs==3.0.2
django-templated-mail==1.0.0
djangorestframework==3.12.4
django-redis==4.12.1
django-redis-sessions==0.6.2
gevent==21.1.2
greenlet==1.0.0
kombu==5.1.0
pika==0.11.2
pytz==2017.3
simplejson==3.13.2
uWSGI==2.0.19.1
vine==5.0.0
dj-database-url==0.5.0
whitenoise==5.2.0
dnspython==2.1.0
six
ipdb
sentry-sdk